+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Create minimal version without styles](https://wealize.atlassian.net/browse/WLZAA-122)
+
+
 ## [v2021.01.29]
 
 - [Move static files (images/videos) to S3 bucket](https://wealize.atlassian.net/browse/WLZAA-119)